+++ lib/rails_health_checks/check_registry.rb
@@ -9,6 +9,10 @@ module RailsHealthChecks
       database: -> { Checks::DatabaseCheck.new },
       cache:    -> { Checks::CacheCheck.new },
       redis:    -> { Checks::RedisCheck.new(url: RailsHealthChecks.configuration.redis_url) },
+      smtp:     -> { Checks::SmtpCheck.new(
+        address: RailsHealthChecks.configuration.smtp_address,
+        port:    RailsHealthChecks.configuration.smtp_port
+      ) },
       sidekiq:     -> { Checks::SidekiqCheck.new(queue_size: RailsHealthChecks.configuration.sidekiq_queue_size) },
       solid_queue: -> { Checks::SolidQueueCheck.new(job_count: RailsHealthChecks.configuration.solid_queue_job_count) },
       good_job:    -> { Checks::GoodJobCheck.new(latency: RailsHealthChecks.configuration.good_job_latency) },